mimo_ofdm_sync: Rewrite state machine of block, simplifying the data flow by adjusting the buffer size to the packet length.

+++ gr-digital/include/gnuradio/digital/mimo_ofdm_synchronizer_fbcvc.h
@@ -48,7 +48,7 @@ namespace gr {
        * class. digital::mimo_ofdm_synchronizer_fbcvc::make is the public interface for
        * creating new instances.
        */
-      static sptr make(uint16_t n, uint16_t fft_len, uint16_t cp_len, const std::vector<gr_complex> &sync_symbol1, const std::vector<gr_complex> &sync_symbol2);
+      static sptr make(uint16_t n, uint32_t fft_len, uint32_t cp_len, const std::vector<gr_complex> &sync_symbol1, const std::vector<gr_complex> &sync_symbol2);
     };
 
   } // namespace digital
